+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   defined in their environment.
 - Warn before sending emails that may need an attachment with
   `no-attachment-warning` in `aerc.conf`.
+- 3 panel view via `:split` and `:vsplit`
 
 ### Changed